refactor: separate user login into layered architecture
