+++ backend/src/routes.ts
@@ -1,10 +1,12 @@
-import { Router } from 'express';
-import { register, login, googleLogin } from './controllers/authController'; 
+import { Router } from "express";
+import { register, googleLogin } from "./controllers/authController";
+import { userLoginRouter } from "./routes/user_login-routes";
 
 const router = Router();
 
-router.post('/register', register);
-router.post('/login', login);
-router.post('/auth/google', googleLogin);
+router.post("/register", register);
+router.post("/auth/google", googleLogin);
+
+router.use(userLoginRouter);
 
 export default router;
