+++ backend/src/routes/user_login-routes.ts
@@ -1,0 +1,6 @@
+import { Router } from "express";
+import { login } from "../controllers/user_login-controller";
+
+export const userLoginRouter = Router();
+
+userLoginRouter.post("/login", login);
